requirements.txt
fastapi
uvicorn
PyPDF2
python-docx
google-generativeai
python-multipart
python-dotenv
openpyxl
